Moving tests, objects and classes; done some fixes to tests
