+++ package.json
@@ -11,7 +11,7 @@
     "start": "node server/build/index.js",
     "test": "node test.js"
   },
-  "type" : "module",
+  "type": "module",
   "dependencies": {
     "@types/express": "^4.17.13",
     "@types/node": "^17.0.33",
@@ -19,12 +19,14 @@
     "bufferutil": "^4.0.6",
     "ejs": "^3.1.5",
     "express": "^4.18.1",
+    "jest": "^28.1.0",
     "ts-node": "^10.7.0",
     "typescript": "^4.6.4",
     "utf-8-validate": "^5.0.9",
     "ws": "^8.6.0"
   },
   "devDependencies": {
+    "@types/jest": "^27.5.1",
     "got": "^11.3.0",
     "tape": "^4.7.0"
   },
