+++ server/src/handler/RevealCardsHandler.ts
