+++ server/src/objects/Fibonacci.ts
@@ -1,0 +1,12 @@
+export enum Fibonacci {
+    A = '1',
+    B = '2',
+    C = '3',
+    D = '5',
+    E = '8',
+    F = '13',
+    G = '21',
+    H = '34',
+    I = '55',
+    J = '89',
+}
