+++ server/src/objects/Primes.ts
@@ -1,0 +1,14 @@
+export enum Primes {
+    A = '2',
+    B = '3',
+    C = '5',
+    D = '7',
+    E = '11',
+    F = '13',
+    G = '17',
+    H = '19',
+    I = '23',
+    J = '29',
+    K = '31',
+    L = '37',
+}
