+++ server/src/objects/TShirts.ts
@@ -1,0 +1,9 @@
+export enum TShirts {
+    A = 'XS',
+    B = 'S',
+    C = 'M',
+    D = 'L',
+    E = 'XL',
+    F = 'XXL',
+    G = 'XXL',
+}
