+++ server/src/tests/NewRoundHandler.test.ts
@@ -1,12 +1,12 @@
-const {NewRoundHandler} = require('./handler/NewRoundHandler');
+import { NewRoundHandler } from '../handler/NewRoundHandler';
 
 describe('ChooseCardHandler', () => {
-    const newRoundMessage = {
+    const newRoundMessage: JSON = JSON.parse(`{
         method: 'newRound',
         params: {
             gameId: '9ef90a72-c735-475a-92c9-0e1cabe63926',
         }
-    };
+    }`);
 
     it('ChooseCardHandler', async () => {
         const handler = new NewRoundHandler();
