+++ server/tsconfig.json
@@ -21,6 +21,7 @@
     "noEmit": false
   },
   "include": [
-    "src/*"
+    "src/*",
+    "src/**/*"
   ]
 }
